Final tests

+++ torch_points3d/utils/box_utils.py
@@ -11,7 +11,7 @@ def box_corners_from_param(box_size, heading_angle, center):
         output (8,3) array for 3D box corners
     """
     R = euler_angles_to_rotation_matrix(torch.tensor([0.0, 0.0, float(heading_angle)]))
-    l, w, h = box_size.float()
+    l, w, h = box_size
     x_corners = torch.tensor([-l / 2, l / 2, l / 2, -l / 2, -l / 2, l / 2, l / 2, -l / 2])
     y_corners = torch.tensor([-w / 2, -w / 2, w / 2, w / 2, -w / 2, -w / 2, w / 2, w / 2])
     z_corners = torch.tensor([-h / 2, -h / 2, -h / 2, -h / 2, h / 2, h / 2, h / 2, h / 2])
